L1_cloud: read max view. zen. angle from ref data
Minor change to make the code read the maximum viewing zenith angle
from the raman effect reference data table. Also fixed a format bug in
m_initialize, and added an if statement in the logging output of
m_get_ai_refl, to avoid trying to write out uninitialized values
if the subroutine is called but cloud fraction is not calculated.