L1_trace_gas & libmet: Read synthetic meteorology files
Modifications to L1_trace_gas to allow tropopause pressure to be read
in from a meteorology file. Meteorology file is handled by a new library,
libmet, which will transparently handle netCDF synthetic and GRIB2 real
meteorology files, but for now only reads netCDF. Alterations to Makefiles
to support building of libmet and a basic unit test of the netCDF read
capability.