L1_cloud: avoid legacy OMI variables in TEMPO mode
When using OMI data, the MeasurementQualityFlag and XtrackQualityFlag
are read from the data and MQF is written in the output. To avoid the
need to include these in TEMPO data, introduced a TEMPO mode switch.
When set, code does not attempt to read/write these flags, and sets
their value to zero.