L1_trace_gas: infrastructure for super-Gaussian
Updated the code to read in the super-Gaussian shape parameter k and
include it in the fitting process when required. Passing k=0 into
super-Gaussian equation will cause a divide-by-zero error, so added
two checks: a) if input k values allow k=0, output a warning to the
log, and b) if k<1e-8 input to super-G module, cancel the convolution
and return (as is the case when half-width = 0).